+++ Guilherme_Ricardo_Dos_Santos_Perinotti/segundo bimestre/recharge-api/src/main/java/br/com/fag/infra/celcoin/dto/response/CelcoinRechargeReceiptDTO.java
@@ -1,4 +1,4 @@
-package br.com.fag.infra.celcoin.dto;
+package br.com.fag.infra.celcoin.dto.response;
 
 import lombok.Getter;
 import lombok.Setter;
