+++ Guilherme_Ricardo_Dos_Santos_Perinotti/segundo bimestre/recharge-api/src/main/java/br/com/fag/infra/celcoin/dto/response/CelcoinTokenDTO.java
@@ -1,4 +1,4 @@
-package br.com.fag.infra.celcoin.dto;
+package br.com.fag.infra.celcoin.dto.response;
 
 import jakarta.json.bind.annotation.JsonbProperty;
 import lombok.Getter;
