+++ Guilherme_Ricardo_Dos_Santos_Perinotti/segundo bimestre/recharge-api/src/main/java/br/com/fag/infra/celcoin/mappers/CelcoinOperatorMapper.java
@@ -1,21 +1,21 @@
 package br.com.fag.infra.celcoin.mappers;
 
 import br.com.fag.domain.dto.OperatorDTO;
-import br.com.fag.infra.celcoin.dto.CelcoinOperatorDTO;
+import br.com.fag.infra.celcoin.dto.response.CelcoinOperatorDTO;
 
 public class CelcoinOperatorMapper {
 
-    public static OperatorDTO toAppDTO(CelcoinOperatorDTO vendorDTO) {
-        OperatorDTO appDTO = new OperatorDTO();
+  public static OperatorDTO toAppDTO(CelcoinOperatorDTO vendorDTO) {
+    OperatorDTO appDTO = new OperatorDTO();
 
-        appDTO.setCategory(vendorDTO.getCategory());
-        appDTO.setMaxValue(vendorDTO.getMaxValue());
-        appDTO.setMinValue(vendorDTO.getMinValue());
-        appDTO.setName(vendorDTO.getName());
-        appDTO.setProviderId(vendorDTO.getProviderId());
-        appDTO.setRechargeType(vendorDTO.getRechargeType());
+    appDTO.setCategory(vendorDTO.getCategory());
+    appDTO.setMaxValue(vendorDTO.getMaxValue());
+    appDTO.setMinValue(vendorDTO.getMinValue());
+    appDTO.setName(vendorDTO.getName());
+    appDTO.setProviderId(vendorDTO.getProviderId());
+    appDTO.setRechargeType(vendorDTO.getRechargeType());
 
-        return appDTO;
-    }
+    return appDTO;
+  }
 
 }
